move the opportunities de-duplication into its own module
Its test imported @/server/opportunities, which builds the auth instance at
import time and throws without APP_URL — so the suite passed locally off a
.env and failed in CI.
