+++ apps/web/src/server/__tests__/opportunities-dedupe.test.ts
@@ -1,5 +1,6 @@
 import { describe, expect, it } from "vitest";
-import { type OpportunitiesReport, withoutRepeats } from "@/server/opportunities";
+import type { OpportunitiesReport } from "@/server/opportunities";
+import { withoutRepeats } from "@/server/opportunities-dedupe";
 
 function opportunity(title: string, overrides: Partial<OpportunitiesReport["opportunities"][number]> = {}) {
 	return {
